feat: only refetch conversation history when agent is active
